create separate file for chartFunctions and add imports/exports for chart functions
